no extend interface

+++ src/estate/Estate.java
@@ -2,5 +2,5 @@ package estate;
 
 public interface Estate {
     
-    double getPrice();
+    int getPrice();
 }
